Fix offline store ingest query decomposition bug
