--- featurebyte/query_graph/transform/offline_ingest_extractor.py
@@ -1,644 +1,0 @@
-"""
-This module contains offline store ingest query extraction related classes.
-"""
-from typing import Any, Dict, List, Optional, Set, Tuple
-
-from collections import defaultdict
-from dataclasses import dataclass
-
-from featurebyte.enum import DBVarType
-from featurebyte.models.base import PydanticObjectId
-from featurebyte.query_graph.algorithm import dfs_traversal
-from featurebyte.query_graph.enum import GraphNodeType, NodeType
-from featurebyte.query_graph.graph_node.base import GraphNode
-from featurebyte.query_graph.model.entity_relationship_info import (
-    EntityAncestorDescendantMapper,
-    EntityRelationshipInfo,
-)
-from featurebyte.query_graph.model.feature_job_setting import FeatureJobSetting
-from featurebyte.query_graph.model.graph import QueryGraphModel
-from featurebyte.query_graph.node import Node
-from featurebyte.query_graph.node.generic import LookupNode, LookupTargetNode
-from featurebyte.query_graph.node.mixin import AggregationOpStructMixin, BaseGroupbyParameters
-from featurebyte.query_graph.node.nested import (
-    AggregationNodeInfo,
-    OfflineStoreIngestQueryGraphNodeParameters,
-)
-from featurebyte.query_graph.node.request import RequestColumnNode
-from featurebyte.query_graph.transform.base import BaseGraphExtractor
-from featurebyte.query_graph.transform.operation_structure import OperationStructureExtractor
-from featurebyte.query_graph.transform.quick_pruning import QuickGraphStructurePruningTransformer
-
-
-def get_offline_store_table_name(
-    primary_entity_serving_names: List[str],
-    feature_job_setting: Optional[FeatureJobSetting],
-    has_ttl: bool,
-) -> str:
-    """
-    Get offline store table name
-
-    Parameters
-    ----------
-    primary_entity_serving_names: List[str]
-        Primary entity serving names
-    feature_job_setting: Optional[FeatureJobSetting]
-        Feature job setting
-    has_ttl: bool
-        Whether the offline store table has time-to-live property or not
-
-    Returns
-    -------
-    str
-        Offline store table name
-    """
-    entity_part = "_".join(primary_entity_serving_names)
-    table_name = f"fb_entity_{entity_part}"
-    if feature_job_setting:
-        fjs = feature_job_setting.to_seconds()
-        frequency = fjs["frequency"]
-        time_modulo_frequency = fjs["time_modulo_frequency"]
-        blind_spot = fjs["blind_spot"]
-        table_name = f"{table_name}_fjs_{frequency}_{time_modulo_frequency}_{blind_spot}"
-    if has_ttl:
-        table_name = f"{table_name}_ttl"
-    return table_name
-
-
-def extract_dtype_from_graph(
-    graph: QueryGraphModel, output_node: Node, exception_message: Optional[str] = None
-) -> DBVarType:
-    """
-    Extract dtype from the given graph and node name
-
-    Parameters
-    ----------
-    graph: QueryGraphModel
-        Query graph
-    output_node: Node
-        Output node
-    exception_message: Optional[str]
-        Optional exception message to use if the graph has more than one aggregation output
-
-    Returns
-    -------
-    DBVarType
-        DType
-
-    Raises
-    ------
-    ValueError
-        If the graph has more than one aggregation output
-    """
-    op_struct_info = OperationStructureExtractor(graph=graph).extract(
-        node=output_node,
-        keep_all_source_columns=True,
-    )
-    op_struct = op_struct_info.operation_structure_map[output_node.name]
-    if len(op_struct.aggregations) != 1:
-        if exception_message is None:
-            exception_message = "Graph must have exactly one aggregation output"
-        raise ValueError(exception_message)
-    return op_struct.aggregations[0].dtype
-
-
-@dataclass
-class AggregationInfo:
-    """
-    AggregationInfo class stores information about the aggregation-type node. Aggregation-type node
-    includes
-    - GroupByNode (time-to-live)
-    - ItemGroupbyNode (non-time-to-live)
-    - AggregateAsAtNode (non-time-to-live)
-    - LookupNode (SCD lookup, Event lookup & Dimension lookup) (non-time-to-live)
-    - RequestColumnNode
-    """
-
-    agg_node_types: List[NodeType]
-    primary_entity_ids: List[PydanticObjectId]
-    feature_job_settings: List[FeatureJobSetting]
-    request_columns: List[str]
-
-    def __init__(self) -> None:
-        self.agg_node_types = []
-        self.primary_entity_ids = []
-        self.feature_job_settings = []
-        self.request_columns = []
-
-    def __add__(self, other: "AggregationInfo") -> "AggregationInfo":
-        """
-        Add two AggregationInfo objects
-
-        Parameters
-        ----------
-        other: AggregationInfo
-            Other AggregationInfo object
-
-        Returns
-        -------
-        AggregationInfo
-            Added AggregationInfo object
-        """
-        output = AggregationInfo()
-        output.agg_node_types = sorted(self.agg_node_types + other.agg_node_types)
-        output.primary_entity_ids = sorted(set(self.primary_entity_ids + other.primary_entity_ids))
-        output.feature_job_settings = list(
-            set(self.feature_job_settings + other.feature_job_settings)
-        )
-        output.request_columns = sorted(set(self.request_columns + other.request_columns))
-        return output
-
-    @property
-    def has_ttl_agg_type(self) -> bool:
-        """
-        Check whether the aggregation info contains time-to-live aggregation type
-
-        Returns
-        -------
-        bool
-        """
-        return NodeType.GROUPBY in self.agg_node_types
-
-
-@dataclass
-class OfflineStoreIngestQueryGraphGlobalState:  # pylint: disable=too-many-instance-attributes
-    """OfflineStoreIngestQueryGlobalState class"""
-
-    # original feature name
-    feature_name: str
-    # decomposed graph
-    graph: QueryGraphModel
-    # original graph node name to decomposed graph node name mapping
-    node_name_map: Dict[str, str]
-    # entity id to ancestor/descendant mapping
-    entity_ancestor_descendant_mapper: EntityAncestorDescendantMapper
-    # (original graph) node name to aggregation node info mapping (from the original graph)
-    node_name_to_aggregation_info: Dict[str, AggregationInfo]
-    # graph node type counter used to generate non-conflicting feature component suffix
-    graph_node_counter: Dict[GraphNodeType, int]
-    # aggregation node names used to determine whether to start decomposing the graph
-    aggregation_node_names: Set[str]
-    # entity id to serving name mapping
-    entity_id_to_serving_name: Dict[PydanticObjectId, str]
-    # whether the graph is decomposed or not
-    is_decomposed: bool = False
-
-    @classmethod
-    def create(
-        cls,
-        relationships_info: Optional[List[EntityRelationshipInfo]],
-        feature_name: str,
-        aggregation_node_names: Set[str],
-        entity_id_to_serving_name: Dict[PydanticObjectId, str],
-    ) -> "OfflineStoreIngestQueryGraphGlobalState":
-        """
-        Create a new OfflineStoreIngestQueryGlobalState object from the given relationships info
-
-        Parameters
-        ----------
-        relationships_info: Optional[List[EntityRelationshipInfo]]
-            Entity relationship info
-        feature_name: str
-            Feature name
-        aggregation_node_names: Set[str]
-            Aggregation node names
-        entity_id_to_serving_name: Dict[PydanticObjectId, str]
-            Entity id to serving name mapping
-
-        Returns
-        -------
-        OfflineStoreIngestQueryGraphGlobalState
-        """
-        return OfflineStoreIngestQueryGraphGlobalState(
-            feature_name=feature_name,
-            graph=QueryGraphModel(),
-            entity_ancestor_descendant_mapper=EntityAncestorDescendantMapper.create(
-                relationships_info=relationships_info or []
-            ),
-            node_name_to_aggregation_info={},
-            graph_node_counter=defaultdict(int),
-            node_name_map={},
-            aggregation_node_names=aggregation_node_names,
-            entity_id_to_serving_name=entity_id_to_serving_name,
-        )
-
-    def update_aggregation_info(self, node: Node, input_node_names: List[str]) -> None:
-        """
-        Update aggregation info of the given node
-
-        Parameters
-        ----------
-        node: Node
-            Node to be processed
-        input_node_names: List[str]
-            List of input node names
-        """
-        aggregation_info = AggregationInfo()
-        for input_node_name in input_node_names:
-            input_aggregation_info = self.node_name_to_aggregation_info[input_node_name]
-            aggregation_info += input_aggregation_info
-
-        if node.name in self.aggregation_node_names:
-            aggregation_info.agg_node_types = [node.type]
-
-        if isinstance(node.parameters, BaseGroupbyParameters):
-            # primary entity ids introduced by groupby node family
-            aggregation_info.primary_entity_ids = node.parameters.entity_ids or []
-        elif isinstance(node, (LookupNode, LookupTargetNode)):
-            # primary entity ids introduced by lookup node family
-            aggregation_info.primary_entity_ids = [node.parameters.entity_id]
-
-        if isinstance(node, RequestColumnNode):
-            # request columns introduced by request column node
-            aggregation_info.request_columns = [node.parameters.column_name]
-
-        if isinstance(node, AggregationOpStructMixin):
-            feature_job_setting = node.extract_feature_job_setting()
-            if feature_job_setting:
-                # feature job settings introduced by aggregation-type node
-                aggregation_info.feature_job_settings = [feature_job_setting]
-
-        # reduce the primary entity ids based on entity relationship
-        aggregation_info.primary_entity_ids = self.entity_ancestor_descendant_mapper.reduce_entity_ids(
-            entity_ids=aggregation_info.primary_entity_ids  # type: ignore
-        )
-
-        # update the mapping
-        self.node_name_to_aggregation_info[node.name] = aggregation_info
-
-    def should_decompose_query_graph(self, node_name: str, input_node_names: List[str]) -> bool:
-        """
-        Check whether to decompose the query graph into graph with nested offline store ingest query nodes
-
-        Parameters
-        ----------
-        node_name: str
-            Node name
-        input_node_names: List[str]
-            List of input node names
-
-        Returns
-        -------
-        bool
-        """
-        aggregation_info = self.node_name_to_aggregation_info[node_name]
-        if not aggregation_info.agg_node_types:
-            # do not decompose if aggregation operation has not been introduced
-            return False
-
-        all_inputs_have_empty_agg_node_types = True
-        for input_node_name in input_node_names:
-            input_agg_info = self.node_name_to_aggregation_info[input_node_name]
-            if (
-                input_agg_info.primary_entity_ids == aggregation_info.primary_entity_ids
-                and bool(input_agg_info.request_columns) == bool(aggregation_info.request_columns)
-                and input_agg_info.feature_job_settings == aggregation_info.feature_job_settings
-                and input_agg_info.has_ttl_agg_type == aggregation_info.has_ttl_agg_type
-            ):
-                # if any of the input is the same as the output, that means
-                # - no new entity ids are added
-                # - no new request columns are added
-                # - no new feature job settings are added
-                # - no new time-to-live aggregation type is added
-                # to the universe.
-                return False
-
-            if input_agg_info.agg_node_types:
-                all_inputs_have_empty_agg_node_types = False
-
-        if all_inputs_have_empty_agg_node_types:
-            # if all the input nodes do not have any aggregation operation introduced, that means
-            # the current node is the first aggregation operation introduced in the graph (between
-            # the input nodes and the current node).
-            return False
-
-        # if none of the above conditions are met, that means we should split the query graph
-        return True
-
-    def add_operation_to_graph(self, node: Node, input_nodes: List[Node]) -> Node:
-        """
-        Add operation to the graph
-
-        Parameters
-        ----------
-        node: Node
-            Node to be added
-        input_nodes: List[Node]
-            List of input nodes
-
-        Returns
-        -------
-        Node
-            Added node
-        """
-        inserted_node = self.graph.add_operation(
-            node_type=node.type,
-            node_params=node.parameters.dict(by_alias=True),
-            node_output_type=node.output_type,
-            input_nodes=input_nodes,
-        )
-        self.node_name_map[node.name] = inserted_node.name
-        return inserted_node
-
-    def get_mapped_decomposed_graph_node(self, node_name: str) -> Node:
-        """
-        Get the mapped decomposed graph node for the given node name
-
-        Parameters
-        ----------
-        node_name: str
-            Node name (of the original graph)
-
-        Returns
-        -------
-        Node
-            Decomposed graph node
-        """
-        return self.graph.get_node_by_name(self.node_name_map[node_name])
-
-
-@dataclass
-class OfflineStoreIngestQueryGraphBranchState:
-    """OfflineStoreIngestQueryBranchState class"""
-
-
-class OfflineStoreIngestQueryGraphExtractor(
-    BaseGraphExtractor[
-        OfflineStoreIngestQueryGraphGlobalState,
-        OfflineStoreIngestQueryGraphBranchState,
-        OfflineStoreIngestQueryGraphGlobalState,
-    ]
-):
-    """
-    OfflineStoreIngestExtractor class
-
-    This class is used to decompose a query graph into a query graph with
-    - offline store ingest query nested graph nodes
-    - post offline store ingest processing nodes
-    """
-
-    def _pre_compute(
-        self,
-        branch_state: OfflineStoreIngestQueryGraphBranchState,
-        global_state: OfflineStoreIngestQueryGraphGlobalState,
-        node: Node,
-        input_node_names: List[str],
-    ) -> Tuple[List[str], bool]:
-        return input_node_names, False
-
-    def _in_compute(
-        self,
-        branch_state: OfflineStoreIngestQueryGraphBranchState,
-        global_state: OfflineStoreIngestQueryGraphGlobalState,
-        node: Node,
-        input_node: Node,
-    ) -> OfflineStoreIngestQueryGraphBranchState:
-        return branch_state
-
-    @staticmethod
-    def _prepare_offline_store_ingest_query_specific_node_parameters(
-        subgraph: QueryGraphModel,
-        subgraph_output_node: Node,
-        node_name_to_subgraph_node_name: Dict[str, str],
-        aggregation_node_names: Set[str],
-        aggregation_info: AggregationInfo,
-        entity_id_to_serving_name: Dict[PydanticObjectId, str],
-    ) -> Dict[str, Any]:
-        """
-        Prepare offline store ingest query graph specific node parameters
-
-        Parameters
-        ----------
-        subgraph: QueryGraphModel
-            Subgraph of the original graph (used to create the offline store ingest query)
-        subgraph_output_node: Node
-            Subgraph output node
-        node_name_to_subgraph_node_name: Dict[str, str]
-            Original graph node name to subgraph node name mapping
-        aggregation_node_names: Set[str]
-            Aggregation node names from the original graph
-        aggregation_info: AggregationInfo
-            Aggregation info of the current node
-        entity_id_to_serving_name: Dict[PydanticObjectId, str]
-            Primary entity id to serving name mapping
-
-        Returns
-        -------
-        Dict[str, Any]
-            Offline store ingest query graph node parameters
-        """
-        agg_nodes_info = []
-        feature_job_settings = []
-        for node_name in aggregation_node_names:
-            if node_name in node_name_to_subgraph_node_name:
-                # if the aggregation node is in the subgraph, that means the aggregation node
-                # is used to create the offline store ingest query
-                subgraph_agg_node_name = node_name_to_subgraph_node_name[node_name]
-                subgraph_agg_node = subgraph.get_node_by_name(subgraph_agg_node_name)
-                input_node_names = subgraph.get_input_node_names(subgraph_agg_node)
-                assert (
-                    len(input_node_names) == 1
-                ), "All non-request column agg. nodes expect only 1 input node"
-                agg_nodes_info.append(
-                    AggregationNodeInfo(
-                        node_type=subgraph_agg_node.type,
-                        node_name=subgraph_agg_node_name,
-                        input_node_name=input_node_names[0],
-                    )
-                )
-
-                if isinstance(subgraph_agg_node, AggregationOpStructMixin):
-                    feature_job_setting = subgraph_agg_node.extract_feature_job_setting()
-                    if feature_job_setting:
-                        feature_job_settings.append(feature_job_setting)
-
-        assert len(set(feature_job_settings)) <= 1, "Only 1 feature job setting is allowed"
-        primary_entity_serving_names = [
-            entity_id_to_serving_name.get(entity_id, str(entity_id))
-            for entity_id in aggregation_info.primary_entity_ids
-        ]
-        feature_job_setting = feature_job_settings[0] if feature_job_settings else None
-        offline_store_table_name = get_offline_store_table_name(
-            primary_entity_serving_names=primary_entity_serving_names,
-            feature_job_setting=feature_job_setting,
-            has_ttl=aggregation_info.has_ttl_agg_type,
-        )
-        output_dtype = extract_dtype_from_graph(graph=subgraph, output_node=subgraph_output_node)
-        parameters = {
-            "aggregation_nodes_info": agg_nodes_info,
-            "feature_job_setting": feature_job_setting,
-            "has_ttl": aggregation_info.has_ttl_agg_type,
-            "offline_store_table_name": offline_store_table_name,
-            "output_dtype": output_dtype,
-        }
-        return parameters
-
-    def _insert_offline_store_query_graph_node(
-        self, global_state: OfflineStoreIngestQueryGraphGlobalState, node_name: str
-    ) -> Node:
-        """
-        Insert offline store ingest query node to the decomposed graph
-
-        Parameters
-        ----------
-        global_state: OfflineStoreIngestQueryGraphGlobalState
-            OfflineStoreIngestQueryGlobalState object
-        node_name: str
-            Node name of the original graph that is used to create the offline store ingest query node
-
-        Returns
-        -------
-        Node
-            Added node (of the decomposed graph)
-        """
-        transformer = QuickGraphStructurePruningTransformer(graph=self.graph)
-        subgraph, node_name_to_transformed_node_name = transformer.transform(
-            target_node_names=[node_name]
-        )
-        transformed_node = subgraph.get_node_by_name(node_name_to_transformed_node_name[node_name])
-        aggregation_info = global_state.node_name_to_aggregation_info[node_name]
-        other_params = self._prepare_offline_store_ingest_query_specific_node_parameters(
-            subgraph=subgraph,
-            subgraph_output_node=transformed_node,
-            node_name_to_subgraph_node_name=node_name_to_transformed_node_name,
-            aggregation_node_names=global_state.aggregation_node_names,
-            aggregation_info=aggregation_info,
-            entity_id_to_serving_name=global_state.entity_id_to_serving_name,
-        )
-        graph_node_type = GraphNodeType.OFFLINE_STORE_INGEST_QUERY
-        part_num = global_state.graph_node_counter[graph_node_type]
-        column_name = f"__{global_state.feature_name}__part{part_num}"
-        graph_node = GraphNode(
-            name="graph",
-            output_type=transformed_node.output_type,
-            parameters=OfflineStoreIngestQueryGraphNodeParameters(
-                graph=subgraph,
-                output_node_name=transformed_node.name,
-                output_column_name=column_name,
-                primary_entity_ids=aggregation_info.primary_entity_ids,
-                **other_params,
-            ),
-        )
-        inserted_node = global_state.add_operation_to_graph(node=graph_node, input_nodes=[])
-
-        # update graph node type counter
-        global_state.graph_node_counter[graph_node_type] += 1
-        return inserted_node
-
-    def _insert_request_column_subgraph(
-        self, global_state: OfflineStoreIngestQueryGraphGlobalState, node_name: str
-    ) -> Node:
-        node_names_to_insert = set()
-        for node in dfs_traversal(
-            query_graph=self.graph, node=self.graph.get_node_by_name(node_name)
-        ):
-            node_names_to_insert.add(node.name)
-
-        for node in self.graph.iterate_sorted_nodes():
-            if node.name in node_names_to_insert:
-                input_nodes = [
-                    global_state.graph.get_node_by_name(global_state.node_name_map[in_node_name])
-                    for in_node_name in self.graph.get_input_node_names(node=node)
-                ]
-                global_state.add_operation_to_graph(node=node, input_nodes=input_nodes)
-
-        mapped_node_name = global_state.node_name_map[node_name]
-        return global_state.graph.get_node_by_name(mapped_node_name)
-
-    def _insert_into_graph(
-        self, global_state: OfflineStoreIngestQueryGraphGlobalState, node_name: str
-    ) -> Node:
-        aggregation_info = global_state.node_name_to_aggregation_info[node_name]
-        if aggregation_info.request_columns:
-            return self._insert_request_column_subgraph(
-                global_state=global_state, node_name=node_name
-            )
-
-        return self._insert_offline_store_query_graph_node(
-            global_state=global_state, node_name=node_name
-        )
-
-    def _post_compute(
-        self,
-        branch_state: OfflineStoreIngestQueryGraphBranchState,
-        global_state: OfflineStoreIngestQueryGraphGlobalState,
-        node: Node,
-        inputs: List[Any],
-        skip_post: bool,
-    ) -> None:
-        input_node_names = self.graph.get_input_node_names(node)
-        global_state.update_aggregation_info(node=node, input_node_names=input_node_names)
-
-        if not global_state.is_decomposed:
-            # check whether to decompose the query graph
-            to_decompose = global_state.should_decompose_query_graph(
-                node_name=node.name, input_node_names=input_node_names
-            )
-            if to_decompose:
-                # insert offline store ingest query node
-                decom_input_nodes = []
-                for input_node_name in input_node_names:
-                    added_node = self._insert_into_graph(
-                        global_state=global_state, node_name=input_node_name
-                    )
-                    decom_input_nodes.append(added_node)
-
-                # add current node to the decomposed graph
-                global_state.add_operation_to_graph(node=node, input_nodes=decom_input_nodes)
-
-            # update global state
-            global_state.is_decomposed = to_decompose
-        else:
-            # if the graph is already decided to be decomposed
-            # first, check if any of the input node has its corresponding mapping in the decomposed graph
-            has_input_associated_with_decomposed_graph = any(
-                input_node_name in global_state.node_name_map
-                for input_node_name in input_node_names
-            )
-            if has_input_associated_with_decomposed_graph:
-                # if any of the input node has its corresponding mapping in the decomposed graph
-                # that means we should insert the current node to the decomposed graph.
-                decom_input_nodes = []
-                for input_node_name in input_node_names:
-                    if input_node_name in global_state.node_name_map:
-                        decom_input_nodes.append(
-                            global_state.get_mapped_decomposed_graph_node(node_name=input_node_name)
-                        )
-                    else:
-                        decom_input_nodes.append(
-                            self._insert_into_graph(
-                                global_state=global_state, node_name=input_node_name
-                            )
-                        )
-
-                # add current node to the decomposed graph
-                global_state.add_operation_to_graph(node=node, input_nodes=decom_input_nodes)
-
-    def extract(
-        self,
-        node: Node,
-        entity_id_to_serving_name: Optional[Dict[PydanticObjectId, str]] = None,
-        relationships_info: Optional[List[EntityRelationshipInfo]] = None,
-        feature_name: str = "feature",
-        **kwargs: Any,
-    ) -> OfflineStoreIngestQueryGraphGlobalState:
-        # identify aggregation node names in the graph, aggregation node names are used to determine
-        # whether to start decomposing the graph
-        op_struct_state = OperationStructureExtractor(graph=self.graph).extract(node=node)
-        op_struct = op_struct_state.operation_structure_map[node.name]
-        aggregation_node_names = {agg.node_name for agg in op_struct.iterate_aggregations()}
-
-        # create global state
-        global_state = OfflineStoreIngestQueryGraphGlobalState.create(
-            relationships_info=relationships_info,
-            feature_name=feature_name,
-            aggregation_node_names=aggregation_node_names,
-            entity_id_to_serving_name=entity_id_to_serving_name or {},
-        )
-        self._extract(
-            node=node,
-            branch_state=OfflineStoreIngestQueryGraphBranchState(),
-            global_state=global_state,
-            topological_order_map=self.graph.node_topological_order_map,
-        )
-        return global_state
